+++ src/io/writer.cpp
@@ -59,62 +59,62 @@ void ThreadedIO::wait() {
     }
 }
 void ThreadedIO::create_file() {
-    if (gsl::at(status, static_cast<size_t>(current_thread)) == WriterStatus::WRITING) {
+    if (gsl::at(status, static_cast<long>(current_thread)) == WriterStatus::WRITING) {
         if (writer_owner == this) {
             wait();
         }
-        gsl::at(status, static_cast<size_t>(current_thread)) = WriterStatus::IDLE;
+        gsl::at(status, static_cast<long>(current_thread)) = WriterStatus::IDLE;
     }
-    if (gsl::at(status, static_cast<size_t>(current_thread)) == WriterStatus::APPENDING) {
+    if (gsl::at(status, static_cast<long>(current_thread)) == WriterStatus::APPENDING) {
         std::cout << "And what should happen to the old file ?" << std::endl;
         exit(ErrorCodes::IO_ERROR);
     }
     // Prepare Xmf file
-    gsl::at(xmf_file, static_cast<size_t>(current_thread)) = XdmfDomain::New();
+    gsl::at(xmf_file, static_cast<long>(current_thread)) = XdmfDomain::New();
     shared_ptr<XdmfInformation> xmf_info = XdmfInformation::New("Copyright", "Produced by MCAC");
-    gsl::at(xmf_file, static_cast<size_t>(current_thread))->insert(xmf_info);
+    gsl::at(xmf_file, static_cast<long>(current_thread))->insert(xmf_info);
 
     // Save simulation properties in every xmf file
-    gsl::at(xmf_file, static_cast<size_t>(current_thread))->insert(physicalmodel->xmf_write());
-    gsl::at(time_collection, static_cast<size_t>(current_thread)) = XdmfGridCollection::New();
-    gsl::at(time_collection, static_cast<size_t>(current_thread))->setType(XdmfGridCollectionType::Temporal());
-    gsl::at(status, static_cast<size_t>(current_thread)) = WriterStatus::APPENDING;
+    gsl::at(xmf_file, static_cast<long>(current_thread))->insert(physicalmodel->xmf_write());
+    gsl::at(time_collection, static_cast<long>(current_thread)) = XdmfGridCollection::New();
+    gsl::at(time_collection, static_cast<long>(current_thread))->setType(XdmfGridCollectionType::Temporal());
+    gsl::at(status, static_cast<long>(current_thread)) = WriterStatus::APPENDING;
 }
 void ThreadedIO::write(const fs::path &prefix, const shared_ptr<XdmfUnstructuredGrid> &data, bool all) {
     if (step % _n_time_per_file == 0) {
         create_file();
     }
-    if (gsl::at(status, static_cast<size_t>(current_thread)) != WriterStatus::APPENDING) {
+    if (gsl::at(status, static_cast<long>(current_thread)) != WriterStatus::APPENDING) {
         std::cout << "File is not ready..." << std::endl;
         exit(ErrorCodes::IO_ERROR);
     }
-    gsl::at(time_collection, static_cast<size_t>(current_thread))->insert(data);
+    gsl::at(time_collection, static_cast<long>(current_thread))->insert(data);
     step++;
     if (step % _n_time_per_file == 0 || all) {
-        gsl::at(xmf_file, static_cast<size_t>(current_thread))
-            ->insert(gsl::at(time_collection, static_cast<size_t>(current_thread)));
+        gsl::at(xmf_file, static_cast<long>(current_thread))
+            ->insert(gsl::at(time_collection, static_cast<long>(current_thread)));
         std::string file_name = std::string(prefix) + filename(num_file, _n);
         wait();
 
         //std::cout << "Writing " << file_name << std::endl;
         if (true) {
             writer =
-                new std::thread(write_task, file_name, &gsl::at(xmf_file, static_cast<size_t>(current_thread)));
+                new std::thread(write_task, file_name, &gsl::at(xmf_file, static_cast<long>(current_thread)));
             writer_owner = this;
-            gsl::at(status, static_cast<size_t>(current_thread)) = WriterStatus::WRITING;
+            gsl::at(status, static_cast<long>(current_thread)) = WriterStatus::WRITING;
         } else { // Sequential
-            write_task(file_name, &gsl::at(xmf_file, static_cast<size_t>(current_thread)));
-            gsl::at(status, static_cast<size_t>(current_thread)) = WriterStatus::IDLE;
+            write_task(file_name, &gsl::at(xmf_file, static_cast<long>(current_thread)));
+            gsl::at(status, static_cast<long>(current_thread)) = WriterStatus::IDLE;
         }
-        gsl::at(status, static_cast<size_t>(current_thread)) = WriterStatus::IDLE;
+        gsl::at(status, static_cast<long>(current_thread)) = WriterStatus::IDLE;
         current_thread = !current_thread;
         num_file++;
         if (all) {
-            if (gsl::at(status, static_cast<size_t>(!current_thread)) == WriterStatus::WRITING) {
+            if (gsl::at(status, static_cast<long>(!current_thread)) == WriterStatus::WRITING) {
                 if (ThreadedIO::writer_owner == this) {
                     wait();
                 }
-                gsl::at(status, static_cast<size_t>(!current_thread)) = WriterStatus::IDLE;
+                gsl::at(status, static_cast<long>(!current_thread)) = WriterStatus::IDLE;
             }
             step = 0;
         }
